requirements.txt
cycler==0.10.0
Cython==0.23.4
dask==0.8.0
decorator==4.0.9
matplotlib==1.5.1
networkx==1.11
numpy==1.10.4
Pillow==3.1.1
pyparsing==2.1.0
python-dateutil==2.5.0
pytz==2015.7
scikit-image==0.12.3
scipy==0.17.0
six==1.10.0
tifffile==0.7.0
toolz==0.7.4
